+++ src/common/MainGrid/index.jsx
@@ -6,64 +6,11 @@ import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 import { Chart } from '../'
 import styles from './index.module.css'
-
+import { createPieChartObject, createBarChartObject } from '../../utils'
 function index() {
-    const pieConfig = {
-        type: 'pie',
-        data: {labels: [
-            'Red',
-            'Blue',
-            'Yellow'
-          ],
-          datasets: [{
-            label: 'My First Dataset',
-            data: [300, 50, 100],
-            backgroundColor: [
-              'rgb(255, 99, 132)',
-              'rgb(54, 162, 235)',
-              'rgb(255, 205, 86)'
-            ],
-            hoverOffset: 4
-          }],}
-    }
-
-      const barData = {
-        labels: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul'],
-        datasets: [{
-          label: 'My First Dataset',
-          data: [65, 59, 80, 81, 56, 55, 40],
-          backgroundColor: [
-            'rgba(255, 99, 132, 0.2)',
-            'rgba(255, 159, 64, 0.2)',
-            'rgba(255, 205, 86, 0.2)',
-            'rgba(75, 192, 192, 0.2)',
-            'rgba(54, 162, 235, 0.2)',
-            'rgba(153, 102, 255, 0.2)',
-            'rgba(201, 203, 207, 0.2)'
-          ],
-          borderColor: [
-            'rgb(255, 99, 132)',
-            'rgb(255, 159, 64)',
-            'rgb(255, 205, 86)',
-            'rgb(75, 192, 192)',
-            'rgb(54, 162, 235)',
-            'rgb(153, 102, 255)',
-            'rgb(201, 203, 207)'
-          ],
-          borderWidth: 1
-        }]
-      }
-      const barConfig = {
-        type: 'bar',
-        data: barData,
-        options: {
-          scales: {
-            y: {
-              beginAtZero: true
-            }
-          }
-        },
-      }
+  const pieConfig = createPieChartObject(['one', 'two', 'three', 'four', 'five'], 'my title', [300, 50, 100, 0, 0])
+  const barConfig = createBarChartObject(['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul'], 'another title', [65, 59, 80, 81, 56, 55, 40])
+  
   return (
     <div>
         <Grid container spacing={2} style={{width: '100%'}}>
